Add typings for WebVTT parser.
Add import type declarations.
