+++ src/types/vtt.ts
@@ -1,0 +1,9 @@
+export type VTTCCs = {
+  ccOffset: number;
+  presentationOffset: number;
+  [key: number]: {
+    start: number;
+    prevCC: number;
+    new: boolean;
+  };
+};
